Refs #29991 - Fix deprecation warnings for AuditAssociations
